Move `TiffValue` trait to separate module
